Ensure migration checks if columns exist to help development

+++ app/alembic/versions/96c0c25be6e5_modify_items_table.py
@@ -8,6 +8,12 @@ Create Date: 2023-02-04 10:51:23.781741
 from alembic import op
 import sqlalchemy as sa
 
+def column_exists(table_name, column_name):
+    bind = op.get_context().bind
+    insp = sa.inspect(bind)
+    columns = insp.get_columns(table_name)
+    return any(c["name"] == column_name for c in columns)
+
 
 # revision identifiers, used by Alembic.
 revision = "96c0c25be6e5"
@@ -17,10 +23,14 @@ depends_on = None
 
 
 def upgrade() -> None:
-    op.add_column("items", sa.Column("description", sa.String()))
-    op.add_column("items", sa.Column("order_", sa.Integer()))
+    if not column_exists(table_name="items", column_name="description"):
+        op.add_column("items", sa.Column("description", sa.String()))
+    if not column_exists(table_name="items", column_name="order_"):
+        op.add_column("items", sa.Column("order_", sa.Integer()))
 
 
 def downgrade() -> None:
-    op.remove_column("items", sa.Column("description", sa.String()))
-    op.remove_column("items", sa.Column("order_", sa.Integer()))
+    if column_exists(table_name="items",  column_name="description"):
+        op.remove_column("items", sa.Column("description", sa.String()))
+    if column_exists(table_name="items",  column_name="order_"):
+        op.remove_column("items", sa.Column("order_", sa.Integer()))
